--- src/blog/blog.regex.ts
@@ -1,5 +1,0 @@
-export const blogRegex = {
-  address: /^[a-z][a-z0-9]+$/g,
-  title: /^[가-힣a-zA-Z0-9]+$/,
-  description: /^[가-힣a-zA-Z0-9.]+$/,
-};
