+++ src/blog/dtos/find-first-by-address/request.dto.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty, IsString } from '@inhanbyeol/class-validator';
+
+export class BlogAddressDto {
+  @IsNotEmpty()
+  @IsString()
+  address: string;
+}
