+++ src/blog/dtos/find-many-and-meta-data/request.dto.ts
@@ -1,0 +1,21 @@
+import { PaginationDto } from '../../../_common/_utils/dtos/request.dto';
+import { IBlogFindManyAndMetaData } from './request.interface';
+import { IsIn, IsNotEmpty, IsOptional, IsString } from '@inhanbyeol/class-validator';
+
+export class BlogFindManyAndMetaDataDto extends PaginationDto implements IBlogFindManyAndMetaData {
+  @IsOptional()
+  @IsIn(['title', 'tags', 'nickname'])
+  searchBy?: string;
+
+  @IsOptional()
+  @IsString()
+  searchKeyword?: string;
+
+  @IsNotEmpty()
+  @IsIn(['title'])
+  orderBy: string = 'title';
+
+  @IsNotEmpty()
+  @IsIn(['asc', 'desc'])
+  sortOrder: string = 'desc';
+}
