+++ src/blog/dtos/find-many-and-meta-data/request.interface.ts
@@ -1,0 +1,8 @@
+import { IPagination } from '../../../_common/_utils/interfaces/request.interface';
+
+export interface IBlogFindManyAndMetaData extends IPagination {
+  searchBy?: string;
+  searchKeyword?: string;
+  orderBy: string;
+  sortOrder: string;
+}
